chore(rpc): move scripts to elecord/rpc
